+++ packages/dataviews/src/index.ts
@@ -1,3 +1,5 @@
+export { DataListProps, default as DataList } from "./dataList";
+export { DataRepeaterProps, default as DataRepeater } from "./dataRepeater";
+export { DataTableProps, default as DataTable } from "./dataTable";
 export * from "./dataTypes";
-export { default as DataRepeater, DataRepeaterProps } from "./dataRepeater";
-export { default as DataTable, DataTableProps } from "./dataTable";
+export { default as ResponsiveTable, ResponsiveTableProps } from "./responsiveTable";